extract ActionContentRouter and ReportActionItemFrame
