+++ src/pages/inbox/report/PureReportActionItem.tsx
@@ -13,23 +13,9 @@ import InlineSystemMessage from '@components/InlineSystemMessage';
 import {ModalActions} from '@components/Modal/Global/ModalContext';
 import OfflineWithFeedback from '@components/OfflineWithFeedback';
 import PressableWithSecondaryInteraction from '@components/PressableWithSecondaryInteraction';
-import ReportActionItemEmojiReactions from '@components/Reactions/ReportActionItemEmojiReactions';
-import RenderHTML from '@components/RenderHTML';
 import ChronosOOOListActions from '@components/ReportActionItem/ChronosOOOListActions';
-import CreatedReportForUnapprovedTransactionsAction from '@components/ReportActionItem/CreatedReportForUnapprovedTransactionsAction';
-import CreateHarvestReportAction from '@components/ReportActionItem/CreateHarvestReportAction';
-import ExportIntegration from '@components/ReportActionItem/ExportIntegration';
-import IssueCardMessage from '@components/ReportActionItem/IssueCardMessage';
-import MoneyRequestAction from '@components/ReportActionItem/MoneyRequestAction';
-import MoneyRequestReportPreview from '@components/ReportActionItem/MoneyRequestReportPreview';
-import MovedTransactionAction from '@components/ReportActionItem/MovedTransactionAction';
-import TaskAction from '@components/ReportActionItem/TaskAction';
-import TaskPreview from '@components/ReportActionItem/TaskPreview';
-import TripRoomPreview from '@components/ReportActionItem/TripRoomPreview';
-import UnreportedTransactionAction from '@components/ReportActionItem/UnreportedTransactionAction';
 import {SearchStateContext} from '@components/Search/SearchContext';
 import {useIsOnSearch} from '@components/Search/SearchScopeProvider';
-import {ShowContextMenuActionsContext, ShowContextMenuStateContext} from '@components/ShowContextMenuContext';
 import UnreadActionIndicator from '@components/UnreadActionIndicator';
 import useConfirmModal from '@hooks/useConfirmModal';
 import useLocalize from '@hooks/useLocalize';
@@ -55,51 +41,19 @@ import Permissions from '@libs/Permissions';
 import {getDisplayNameOrDefault} from '@libs/PersonalDetailsUtils';
 import {
     extractLinksFromMessageHtml,
-    getChangedApproverActionMessage,
-    getCompanyCardConnectionBrokenMessage,
-    getIntegrationSyncFailedMessage,
-    getIOUReportIDFromReportActionPreview,
     getOriginalMessage,
-    getPlaidBalanceFailureMessage,
-    getRenamedAction,
-    getReportActionHtml,
     getReportActionMessage,
     getReportActionText,
-    getSettlementAccountLockedMessage,
-    getTravelUpdateMessage,
     getWhisperedTo,
-    isActionableCardFraudAlert,
-    isActionableJoinRequest,
-    isActionableMentionInviteToSubmitExpenseConfirmWhisper,
-    isActionableMentionWhisper,
-    isActionableReportMentionWhisper,
-    isActionOfType,
-    isCardBrokenConnectionAction,
-    isCardIssuedAction,
-    isCreatedTaskReportAction,
     isDeletedAction,
     isDeletedParentAction as isDeletedParentActionUtils,
-    isIOURequestReportAction,
-    isMessageDeleted,
     isMoneyRequestAction,
     isPendingRemove,
-    isReimbursementDeQueuedOrCanceledAction,
-    isReimbursementQueuedAction,
-    isRenamedAction,
-    isTaskAction,
     isTripPreview,
     isWhisperActionTargetedToOthers,
     useTableReportViewActionRenderConditionals,
 } from '@libs/ReportActionsUtils';
-import {
-    canWriteInReport,
-    getMovedActionMessage,
-    isCompletedTaskReport,
-    isExpenseReport,
-    isHarvestCreatedExpenseReport as isHarvestCreatedExpenseReportUtils,
-    isTaskReport,
-    shouldDisplayThreadReplies as shouldDisplayThreadRepliesUtils,
-} from '@libs/ReportUtils';
+import {canWriteInReport, isCompletedTaskReport, isHarvestCreatedExpenseReport as isHarvestCreatedExpenseReportUtils, isTaskReport} from '@libs/ReportUtils';
 import SelectionScraper from '@libs/SelectionScraper';
 import {ReactionListContext} from '@pages/inbox/ReportScreenContext';
 import AttachmentModalContext from '@pages/media/AttachmentModalScreen/AttachmentModalContext';
@@ -111,34 +65,14 @@ import ONYXKEYS from '@src/ONYXKEYS';
 import type SCREENS from '@src/SCREENS';
 import type * as OnyxTypes from '@src/types/onyx';
 import type {Errors} from '@src/types/onyx/OnyxCommon';
-import {isEmptyObject, isEmptyValueObject} from '@src/types/utils/EmptyObject';
-import ApprovalFlowContent, {isApprovalFlowAction} from './actionContents/ApprovalFlowContent';
-import CardBrokenConnectionContent from './actionContents/CardBrokenConnectionContent';
-import ChatMessageContent from './actionContents/ChatMessageContent';
-import ChatTransactionPreview from './actionContents/ChatTransactionPreview';
-import ConfirmWhisperContent from './actionContents/ConfirmWhisperContent';
-import FraudAlertContent from './actionContents/FraudAlertContent';
-import JoinRequestContent from './actionContents/JoinRequestContent';
-import MentionWhisperContent from './actionContents/MentionWhisperContent';
-import ModifiedExpenseContent from './actionContents/ModifiedExpenseContent';
-import PaymentContent from './actionContents/PaymentContent';
-import PolicyChangeLogContent, {isHandledPolicyChangeLogAction} from './actionContents/PolicyChangeLogContent';
-import ReimbursedContent from './actionContents/ReimbursedContent';
-import ReimbursementDeQueuedContent from './actionContents/ReimbursementDeQueuedContent';
-import ReimbursementQueuedContent from './actionContents/ReimbursementQueuedContent';
-import ReportMentionWhisperContent from './actionContents/ReportMentionWhisperContent';
-import SimpleMessageContent, {isSimpleMessageAction} from './actionContents/SimpleMessageContent';
+import {isEmptyValueObject} from '@src/types/utils/EmptyObject';
+import ActionContentRouter from './actionContents/ActionContentRouter';
 import {RestrictedReadOnlyContextMenuActions} from './ContextMenu/ContextMenuActions';
 import MiniReportActionContextMenu from './ContextMenu/MiniReportActionContextMenu';
 import type {ContextMenuAnchor} from './ContextMenu/ReportActionContextMenu';
 import {hideContextMenu, hideDeleteModal, isActiveReportAction, showContextMenu} from './ContextMenu/ReportActionContextMenu';
-import LinkPreviewer from './LinkPreviewer';
-import ReportActionItemBasicMessage from './ReportActionItemBasicMessage';
 import ReportActionItemContentCreated from './ReportActionItemContentCreated';
-import ReportActionItemDraft from './ReportActionItemDraft';
-import ReportActionItemGrouped from './ReportActionItemGrouped';
-import ReportActionItemSingle from './ReportActionItemSingle';
-import ReportActionItemThread from './ReportActionItemThread';
+import ReportActionItemFrame from './ReportActionItemFrame';
 import SearchActionHeader from './SearchActionHeader';
 import TripSummary from './TripSummary';
 import WhisperBanner from './WhisperBanner';
@@ -273,12 +207,6 @@ type PureReportActionItemProps = {
     reportNameValuePairsOriginalID?: string;
 };
 
-// This is equivalent to returning a negative boolean in normal functions, but we can keep the element return type
-// If the child was rendered using RenderHTML and an empty html string, it has an empty prop called html
-// If we render an empty component/fragment, this does not apply
-const emptyHTML = <RenderHTML html="" />;
-const isEmptyHTML = <T extends React.JSX.Element>({props: {html}}: T): boolean => typeof html === 'string' && html.length === 0;
-
 function PureReportActionItem({
     action,
     report,
@@ -321,7 +249,7 @@ function PureReportActionItem({
     const shouldDisplayContextMenuValue = shouldDisplayContextMenu && !isConciergeGreeting;
 
     const {transitionActionSheetState} = ActionSheetAwareScrollView.useActionSheetAwareScrollViewActions();
-    const {translate, formatTravelDate, datetimeToCalendarTime} = useLocalize();
+    const {translate, datetimeToCalendarTime} = useLocalize();
     const {showConfirmModal} = useConfirmModal();
     const {shouldUseNarrowLayout} = useResponsiveLayout();
     const reportID = report?.reportID ?? action?.reportID;
@@ -606,456 +534,22 @@ function PureReportActionItem({
         [toggleContextMenuFromActiveReportAction, handleShowContextMenu],
     );
 
-    /**
-     * Get the content of ReportActionItem
-     * @param hovered whether the ReportActionItem is hovered
-     * @param isWhisper whether the report action is a whisper
-     * @param hasErrors whether the report action has any errors
-     * @returns child component(s)
-     */
-    const renderItemContent = (hovered = false, isWhisper = false, hasErrors = false): React.JSX.Element => {
-        let children;
-        const moneyRequestOriginalMessage = isMoneyRequestAction(action) ? getOriginalMessage(action) : undefined;
-        const moneyRequestActionType = moneyRequestOriginalMessage?.type;
+    const createdTransactionID = isMoneyRequestAction(parentReportActionForTransactionThread) ? getOriginalMessage(parentReportActionForTransactionThread)?.IOUTransactionID : undefined;
 
-        // Show the preview for when expense is present
-        if (isIOURequestReportAction(action)) {
-            const isSplitScanWithNoAmount = moneyRequestActionType === CONST.IOU.REPORT_ACTION_TYPE.SPLIT && moneyRequestOriginalMessage?.amount === 0;
-            const chatReportID = moneyRequestOriginalMessage?.IOUReportID ? report?.chatReportID : reportID;
-            // There is no single iouReport for bill splits, so only 1:1 requests require an iouReportID
-            const iouReportID = moneyRequestOriginalMessage?.IOUReportID?.toString();
-            children = (
-                <MoneyRequestAction
-                    // If originalMessage.iouReportID is set, this is a 1:1 IOU expense in a DM chat whose reportID is report.chatReportID
-                    chatReportID={chatReportID}
-                    requestReportID={iouReportID}
-                    reportID={reportID}
-                    action={action}
-                    isHovered={hovered}
-                    contextMenuAnchor={popoverAnchorRef.current}
-                    checkIfContextMenuActive={toggleContextMenuFromActiveReportAction}
-                    style={displayAsGroup ? [] : [styles.mt2]}
-                    isWhisper={isWhisper}
-                    shouldDisplayContextMenu={shouldDisplayContextMenuValue}
-                    originalReportID={originalReportID}
-                />
-            );
-
-            if (report?.type === CONST.REPORT.TYPE.CHAT) {
-                const isSplitBill = moneyRequestActionType === CONST.IOU.REPORT_ACTION_TYPE.SPLIT;
-                const shouldShowSplitPreview = isSplitBill || isSplitScanWithNoAmount;
-                if (report.chatType === CONST.REPORT.CHAT_TYPE.SELF_DM || shouldShowSplitPreview) {
-                    children = (
-                        <ChatTransactionPreview
-                            action={action}
-                            reportID={reportID}
-                            originalReportID={originalReportID}
-                            chatReportID={chatReportID}
-                            iouReport={iouReport}
-                            shouldShowSplitPreview={shouldShowSplitPreview}
-                            shouldDisplayContextMenu={shouldDisplayContextMenuValue}
-                            transactionID={shouldShowSplitPreview ? moneyRequestOriginalMessage?.IOUTransactionID : undefined}
-                        />
-                    );
-                } else {
-                    children = emptyHTML;
-                }
-            }
-        } else if (isTripPreview(action)) {
-            children = (
-                <TripRoomPreview
-                    action={action}
-                    isHovered={hovered}
-                    contextMenuAnchor={popoverAnchorRef.current}
-                    containerStyles={displayAsGroup ? [] : [styles.mt2]}
-                    checkIfContextMenuActive={toggleContextMenuFromActiveReportAction}
-                    shouldDisplayContextMenu={shouldDisplayContextMenuValue}
-                    originalReportID={originalReportID}
-                />
-            );
-        } else if (action.actionName === CONST.REPORT.ACTIONS.TYPE.REPORT_PREVIEW && isClosedExpenseReportWithNoExpenses) {
-            children = <RenderHTML html={`<deleted-action>${translate('parentReportAction.deletedReport')}</deleted-action>`} />;
-        } else if (action.actionName === CONST.REPORT.ACTIONS.TYPE.REPORT_PREVIEW) {
-            children = (
-                <MoneyRequestReportPreview
-                    iouReportID={getIOUReportIDFromReportActionPreview(action)}
-                    policyID={report?.policyID}
-                    chatReportID={reportID}
-                    action={action}
-                    contextMenuAnchor={popoverAnchorRef.current}
-                    isHovered={hovered}
-                    isWhisper={isWhisper}
-                    checkIfContextMenuActive={toggleContextMenuFromActiveReportAction}
-                    onPaymentOptionsShow={() => setIsPaymentMethodPopoverActive(true)}
-                    onPaymentOptionsHide={() => setIsPaymentMethodPopoverActive(false)}
-                    shouldDisplayContextMenu={shouldDisplayContextMenuValue}
-                    shouldShowBorder={shouldShowBorder}
-                    originalReportID={originalReportID}
-                />
-            );
-        } else if (isTaskAction(action)) {
-            children = <TaskAction action={action} />;
-        } else if (isCreatedTaskReportAction(action)) {
-            children = (
-                <ShowContextMenuStateContext.Provider value={contextMenuStateValue}>
-                    <ShowContextMenuActionsContext.Provider value={contextMenuActionsValue}>
-                        <TaskPreview
-                            style={displayAsGroup ? [] : [styles.mt1]}
-                            chatReportID={reportID}
-                            action={action}
-                            isHovered={hovered}
-                            onShowContextMenu={handleShowContextMenu}
-                            contextMenuAnchor={popoverAnchorRef.current}
-                            checkIfContextMenuActive={toggleContextMenuFromActiveReportAction}
-                            policyID={report?.policyID}
-                            shouldDisplayContextMenu={shouldDisplayContextMenuValue}
-                        />
-                    </ShowContextMenuActionsContext.Provider>
-                </ShowContextMenuStateContext.Provider>
-            );
-        } else if (isReimbursementQueuedAction(action)) {
-            children = (
-                <ReimbursementQueuedContent
-                    action={action}
-                    report={report}
-                    iouReport={iouReport}
-                    personalDetails={personalDetails}
-                />
-            );
-        } else if (isReimbursementDeQueuedOrCanceledAction(action)) {
-            children = (
-                <ReimbursementDeQueuedContent
-                    action={action}
-                    report={report}
-                />
-            );
-        } else if (action.actionName === CONST.REPORT.ACTIONS.TYPE.MODIFIED_EXPENSE) {
-            children = (
-                <ModifiedExpenseContent
-                    action={action}
-                    report={report}
-                    originalReport={originalReport}
-                />
-            );
-        } else if (isApprovalFlowAction(action)) {
-            children = (
-                <ApprovalFlowContent
-                    action={action}
-                    policyID={report?.policyID}
-                    reportID={reportID}
-                    originalReport={originalReport}
-                />
-            );
-        } else if (isActionOfType(action, CONST.REPORT.ACTIONS.TYPE.IOU) && getOriginalMessage(action)?.type === CONST.IOU.REPORT_ACTION_TYPE.PAY) {
-            children = (
-                <PaymentContent
-                    action={action}
-                    policyID={report?.policyID}
-                />
-            );
-        } else if (isActionOfType(action, CONST.REPORT.ACTIONS.TYPE.REIMBURSED)) {
-            children = (
-                <ReimbursedContent
-                    action={action}
-                    report={report}
-                />
-            );
-        } else if (isSimpleMessageAction(action)) {
-            children = (
-                <SimpleMessageContent
-                    action={action}
-                    report={report}
-                />
-            );
-        } else if (isActionOfType(action, CONST.REPORT.ACTIONS.TYPE.FORWARDED)) {
-            const wasAutoForwarded = getOriginalMessage(action)?.automaticAction ?? false;
-            if (wasAutoForwarded) {
-                children = (
-                    <ReportActionItemBasicMessage>
-                        <RenderHTML html={`<comment><muted-text>${translate('iou.automaticallyForwarded')}</muted-text></comment>`} />
-                    </ReportActionItemBasicMessage>
-                );
-            } else {
-                children = <ReportActionItemBasicMessage message={translate('iou.forwarded')} />;
-            }
-        } else if (isHandledPolicyChangeLogAction(action)) {
-            children = (
-                <PolicyChangeLogContent
-                    action={action}
-                    policyID={report?.policyID}
-                />
-            );
-        } else if (action.actionName === CONST.REPORT.ACTIONS.TYPE.MOVED_TRANSACTION) {
-            children = (
-                <MovedTransactionAction
-                    action={action as OnyxTypes.ReportAction<typeof CONST.REPORT.ACTIONS.TYPE.MOVED_TRANSACTION>}
-                    emptyHTML={emptyHTML}
-                    originalReport={originalReport}
-                />
-            );
-        } else if (action.actionName === CONST.REPORT.ACTIONS.TYPE.MOVED) {
-            children = (
-                <ReportActionItemBasicMessage message="">
-                    <RenderHTML html={`<comment><muted-text>${getMovedActionMessage(translate, action, report)}</muted-text></comment>`} />
-                </ReportActionItemBasicMessage>
-            );
-        } else if (isActionOfType(action, CONST.REPORT.ACTIONS.TYPE.TRAVEL_UPDATE)) {
-            children = (
-                <ReportActionItemBasicMessage message="">
-                    <RenderHTML html={`<comment><muted-text>${getTravelUpdateMessage(translate, action, formatTravelDate)}</muted-text></comment>`} />
-                </ReportActionItemBasicMessage>
-            );
-        } else if (action.actionName === CONST.REPORT.ACTIONS.TYPE.UNREPORTED_TRANSACTION) {
-            children = (
-                <UnreportedTransactionAction
-                    action={action as OnyxTypes.ReportAction<typeof CONST.REPORT.ACTIONS.TYPE.UNREPORTED_TRANSACTION>}
-                    originalReport={originalReport}
-                />
-            );
-        } else if (action.actionName === CONST.REPORT.ACTIONS.TYPE.CARD_FROZEN || action.actionName === CONST.REPORT.ACTIONS.TYPE.CARD_UNFROZEN) {
-            children = (
-                <ReportActionItemBasicMessage message="">
-                    <RenderHTML html={`<comment><muted-text>${getReportActionHtml(action)}</muted-text></comment>`} />
-                </ReportActionItemBasicMessage>
-            );
-        } else if (isActionableCardFraudAlert(action)) {
-            children = (
-                <FraudAlertContent
-                    action={action}
-                    reportID={reportID}
-                />
-            );
-        } else if (isActionableJoinRequest(action)) {
-            children = (
-                <JoinRequestContent
-                    action={action}
-                    reportID={reportID}
-                    originalReportID={originalReportID}
-                    policyID={report?.policyID}
-                />
-            );
-        } else if (isActionableMentionWhisper(action)) {
-            children = (
-                <MentionWhisperContent
-                    action={action}
-                    report={report}
-                    originalReport={originalReport}
-                    originalReportID={originalReportID}
-                    resolveActionableMentionWhisper={resolveActionableMentionWhisper}
-                />
-            );
-        } else if (isActionableReportMentionWhisper(action)) {
-            children = (
-                <ReportMentionWhisperContent
-                    action={action}
-                    reportID={reportID}
-                    report={report}
-                    originalReport={originalReport}
-                    isReportArchived={isReportArchived}
-                    resolveActionableReportMentionWhisper={resolveActionableReportMentionWhisper}
-                />
-            );
-        } else if (isActionableMentionInviteToSubmitExpenseConfirmWhisper(action)) {
-            children = (
-                <ConfirmWhisperContent
-                    action={action}
-                    reportID={reportID}
-                    report={report}
-                    originalReport={originalReport}
-                    originalReportID={originalReportID}
-                />
-            );
-        } else if (isActionOfType(action, CONST.REPORT.ACTIONS.TYPE.ROOM_CHANGE_LOG.LEAVE_ROOM)) {
-            children = <ReportActionItemBasicMessage message={translate('report.actions.type.leftTheChat')} />;
-        } else if (isCardIssuedAction(action)) {
-            children = (
-                <IssueCardMessage
-                    action={action}
-                    policyID={report?.policyID}
-                />
-            );
-        } else if (isCardBrokenConnectionAction(action)) {
-            children = <CardBrokenConnectionContent action={action} />;
-        } else if (isActionOfType(action, CONST.REPORT.ACTIONS.TYPE.EXPORTED_TO_INTEGRATION)) {
-            children = <ExportIntegration action={action} />;
-        } else if (isRenamedAction(action)) {
-            const message = getRenamedAction(translate, action, isExpenseReport(report));
-            children = <ReportActionItemBasicMessage message={message} />;
-        } else if (isActionOfType(action, CONST.REPORT.ACTIONS.TYPE.INTEGRATION_SYNC_FAILED)) {
-            children = (
-                <ReportActionItemBasicMessage message="">
-                    <RenderHTML html={`<comment><muted-text>${getIntegrationSyncFailedMessage(translate, action, report?.policyID, isTryNewDotNVPDismissed)}</muted-text></comment>`} />
-                </ReportActionItemBasicMessage>
-            );
-        } else if (isActionOfType(action, CONST.REPORT.ACTIONS.TYPE.COMPANY_CARD_CONNECTION_BROKEN)) {
-            children = (
-                <ReportActionItemBasicMessage message="">
-                    <RenderHTML html={`<comment><muted-text>${getCompanyCardConnectionBrokenMessage(translate, action)}</muted-text></comment>`} />
-                </ReportActionItemBasicMessage>
-            );
-        } else if (isActionOfType(action, CONST.REPORT.ACTIONS.TYPE.PLAID_BALANCE_FAILURE)) {
-            children = (
-                <ReportActionItemBasicMessage message="">
-                    <RenderHTML html={`<comment><muted-text>${getPlaidBalanceFailureMessage(translate, action)}</muted-text></comment>`} />
-                </ReportActionItemBasicMessage>
-            );
-        } else if (isActionOfType(action, CONST.REPORT.ACTIONS.TYPE.CREATED) && isHarvestCreatedExpenseReport) {
-            children = <CreateHarvestReportAction reportNameValuePairsOriginalID={reportNameValuePairsOriginalID} />;
-        } else if (isActionOfType(action, CONST.REPORT.ACTIONS.TYPE.CREATED_REPORT_FOR_UNAPPROVED_TRANSACTIONS)) {
-            children = <CreatedReportForUnapprovedTransactionsAction action={action} />;
-        } else if (
-            isActionOfType(action, CONST.REPORT.ACTIONS.TYPE.TAKE_CONTROL) ||
-            isActionOfType(action, CONST.REPORT.ACTIONS.TYPE.REROUTE) ||
-            isActionOfType(action, CONST.REPORT.ACTIONS.TYPE.REASSIGN_APPROVER)
-        ) {
-            children = (
-                <ReportActionItemBasicMessage>
-                    <RenderHTML html={`<comment><muted-text>${getChangedApproverActionMessage(translate, action)}</muted-text></comment>`} />
-                </ReportActionItemBasicMessage>
-            );
-        } else if (isActionOfType(action, CONST.REPORT.ACTIONS.TYPE.SETTLEMENT_ACCOUNT_LOCKED)) {
-            children = (
-                <ReportActionItemBasicMessage>
-                    <RenderHTML html={`<comment><muted-text>${getSettlementAccountLockedMessage(translate, action)}</muted-text></comment>`} />
-                </ReportActionItemBasicMessage>
-            );
-        } else {
-            children = (
-                <ChatMessageContent
-                    action={action}
-                    report={report}
-                    originalReport={originalReport}
-                    reportID={reportID}
-                    originalReportID={originalReportID}
-                    displayAsGroup={displayAsGroup}
-                    draftMessage={draftMessage}
-                    index={index}
-                    isHidden={isHidden}
-                    moderationDecision={moderationDecision}
-                    updateHiddenState={updateHiddenState}
-                    isArchivedRoom={isArchivedRoom}
-                    composerTextInputRef={composerTextInputRef}
-                    isOnSearch={isOnSearch}
-                    currentSearchHash={currentSearchHash}
-                    contextMenuStateValue={contextMenuStateValue}
-                    contextMenuActionsValue={contextMenuActionsValue}
-                    userBillingFundID={userBillingFundID}
-                />
-            );
-        }
-        const numberOfThreadReplies = action.childVisibleActionCount ?? 0;
-
-        const shouldDisplayThreadReplies = shouldDisplayThreadRepliesUtils(action, isThreadReportParentAction) && !isOnSearch;
-        const oldestFourAccountIDs =
-            action.childOldestFourAccountIDs
-                ?.split(',')
-                .map((accountID) => Number(accountID))
-                .filter((accountID): accountID is number => typeof accountID === 'number') ?? [];
-        const draftMessageRightAlign = draftMessage !== undefined ? styles.chatItemReactionsDraftRight : {};
-
-        const itemContent = (
-            <>
-                {children}
-                {Permissions.canUseLinkPreviews() && !isHidden && (action.linkMetadata?.length ?? 0) > 0 && (
-                    <View style={draftMessage !== undefined ? styles.chatItemReactionsDraftRight : {}}>
-                        <LinkPreviewer linkMetadata={action.linkMetadata?.filter((item) => !isEmptyObject(item))} />
-                    </View>
-                )}
-                {!isOnSearch && !isMessageDeleted(action) && (
-                    <View style={draftMessageRightAlign}>
-                        <ReportActionItemEmojiReactions
-                            reportAction={action}
-                            reportID={reportID}
-                            shouldBlockReactions={hasErrors}
-                            setIsEmojiPickerActive={setIsEmojiPickerActive}
-                        />
-                    </View>
-                )}
-
-                {shouldDisplayThreadReplies && (
-                    <View style={draftMessageRightAlign}>
-                        <ReportActionItemThread
-                            reportAction={action}
-                            report={report}
-                            numberOfReplies={numberOfThreadReplies}
-                            mostRecentReply={`${action.childLastVisibleActionCreated}`}
-                            isHovered={hovered || isContextMenuActive}
-                            accountIDs={oldestFourAccountIDs}
-                            onSecondaryInteraction={showPopover}
-                            isActive={isReportActionActive && !isContextMenuActive}
-                        />
-                    </View>
-                )}
-            </>
-        );
-
-        return isEmptyHTML(children) ? emptyHTML : itemContent;
-    };
-
-    /**
-     * Get ReportActionItem with a proper wrapper
-     * @param hovered whether the ReportActionItem is hovered
-     * @param isWhisper whether the ReportActionItem is a whisper
-     * @param hasErrors whether the report action has any errors
-     * @returns report action item
-     */
-
-    const renderReportActionItem = (hovered: boolean, isWhisper: boolean, hasErrors: boolean): React.JSX.Element => {
-        const content = renderItemContent(hovered || isContextMenuActive || isEmojiPickerActive, isWhisper, hasErrors);
-
-        if (isEmptyHTML(content) || (!shouldRenderViewBasedOnAction && !isClosedExpenseReportWithNoExpenses)) {
-            return emptyHTML;
-        }
-
-        if (draftMessage !== undefined) {
-            return <ReportActionItemDraft>{content}</ReportActionItemDraft>;
-        }
-
-        if (!displayAsGroup) {
-            return (
-                <ReportActionItemSingle
-                    action={action}
-                    showHeader={draftMessage === undefined}
-                    wrapperStyle={{
-                        ...(isOnSearch && styles.p0),
-                        ...(isWhisper && styles.pt1),
-                    }}
-                    report={report}
-                    iouReport={iouReport}
-                    isHovered={hovered || isContextMenuActive}
-                    isActive={isReportActionActive && !isContextMenuActive}
-                    hasBeenFlagged={
-                        ![CONST.MODERATION.MODERATOR_DECISION_APPROVED, CONST.MODERATION.MODERATOR_DECISION_PENDING].some((item) => item === moderationDecision) && !isPendingRemove(action)
-                    }
-                >
-                    {content}
-                </ReportActionItemSingle>
-            );
-        }
-
-        return <ReportActionItemGrouped wrapperStyle={isWhisper ? styles.pt1 : {}}>{content}</ReportActionItemGrouped>;
-    };
-
-    const createdActionContent = useMemo(() => {
-        const transactionID = isMoneyRequestAction(parentReportActionForTransactionThread) ? getOriginalMessage(parentReportActionForTransactionThread)?.IOUTransactionID : undefined;
-
+    if (action.actionName === CONST.REPORT.ACTIONS.TYPE.CREATED && !isHarvestCreatedExpenseReport) {
         return (
             <ReportActionItemContentCreated
                 contextMenuStateValue={contextMenuStateValue}
                 contextMenuActionsValue={contextMenuActionsValue}
                 parentReportAction={parentReportAction}
-                transactionID={transactionID}
+                transactionID={createdTransactionID}
                 draftMessage={draftMessage}
                 shouldHideThreadDividerLine={shouldHideThreadDividerLine}
             />
         );
-    }, [contextMenuStateValue, contextMenuActionsValue, parentReportAction, draftMessage, shouldHideThreadDividerLine, parentReportActionForTransactionThread]);
-
-    if (action.actionName === CONST.REPORT.ACTIONS.TYPE.CREATED && !isHarvestCreatedExpenseReport) {
-        return createdActionContent;
     }
 
-    // Should show createdActionContent if:
+    // Should show the created action content if:
     // - Harvest report is shown in ReportActionsList (one-transaction view)
     const shouldShowCreatedAction = action.actionName === CONST.REPORT.ACTIONS.TYPE.CREATED && !!isHarvestCreatedExpenseReport && !!parentReportActionForTransactionThread;
 
@@ -1100,7 +594,16 @@ function PureReportActionItem({
 
     return (
         <View>
-            {shouldShowCreatedAction && createdActionContent}
+            {shouldShowCreatedAction && (
+                <ReportActionItemContentCreated
+                    contextMenuStateValue={contextMenuStateValue}
+                    contextMenuActionsValue={contextMenuActionsValue}
+                    parentReportAction={parentReportAction}
+                    transactionID={createdTransactionID}
+                    draftMessage={draftMessage}
+                    shouldHideThreadDividerLine={shouldHideThreadDividerLine}
+                />
+            )}
             <PressableWithSecondaryInteraction
                 ref={popoverAnchorRef}
                 onPress={() => {
@@ -1179,7 +682,65 @@ function PureReportActionItem({
                                         onPress={onPress}
                                     >
                                         {isWhisper && <WhisperBanner whisperedTo={whisperedTo} />}
-                                        {renderReportActionItem(!!hovered || !!isReportActionLinked, isWhisper, hasErrors)}
+                                        <ReportActionItemFrame
+                                            action={action}
+                                            report={report}
+                                            iouReport={iouReport}
+                                            displayAsGroup={displayAsGroup}
+                                            draftMessage={draftMessage}
+                                            isWhisper={isWhisper}
+                                            isOnSearch={isOnSearch}
+                                            hovered={!!hovered || !!isReportActionLinked}
+                                            isContextMenuActive={isContextMenuActive}
+                                            isReportActionActive={isReportActionActive}
+                                            moderationDecision={moderationDecision}
+                                            shouldRenderViewBasedOnAction={shouldRenderViewBasedOnAction}
+                                            isClosedExpenseReportWithNoExpenses={isClosedExpenseReportWithNoExpenses}
+                                        >
+                                            <ActionContentRouter
+                                                action={action}
+                                                report={report}
+                                                originalReport={originalReport}
+                                                originalReportID={originalReportID}
+                                                iouReport={iouReport}
+                                                reportID={reportID}
+                                                displayAsGroup={displayAsGroup}
+                                                draftMessage={draftMessage}
+                                                isWhisper={isWhisper}
+                                                hovered={!!hovered || !!isReportActionLinked || isContextMenuActive || !!isEmojiPickerActive}
+                                                hasErrors={hasErrors}
+                                                isContextMenuActive={isContextMenuActive}
+                                                isReportActionActive={isReportActionActive}
+                                                isHidden={isHidden}
+                                                moderationDecision={moderationDecision}
+                                                updateHiddenState={updateHiddenState}
+                                                isArchivedRoom={isArchivedRoom}
+                                                isReportArchived={isReportArchived}
+                                                isClosedExpenseReportWithNoExpenses={isClosedExpenseReportWithNoExpenses}
+                                                isHarvestCreatedExpenseReport={isHarvestCreatedExpenseReport}
+                                                reportNameValuePairsOriginalID={reportNameValuePairsOriginalID}
+                                                personalDetails={personalDetails}
+                                                isTryNewDotNVPDismissed={isTryNewDotNVPDismissed}
+                                                shouldShowBorder={shouldShowBorder}
+                                                isThreadReportParentAction={isThreadReportParentAction}
+                                                isOnSearch={isOnSearch}
+                                                shouldDisplayContextMenuValue={shouldDisplayContextMenuValue}
+                                                userBillingFundID={userBillingFundID}
+                                                index={index}
+                                                contextMenuAnchor={popoverAnchorRef.current}
+                                                composerTextInputRef={composerTextInputRef}
+                                                contextMenuStateValue={contextMenuStateValue}
+                                                contextMenuActionsValue={contextMenuActionsValue}
+                                                setIsPaymentMethodPopoverActive={setIsPaymentMethodPopoverActive}
+                                                setIsEmojiPickerActive={setIsEmojiPickerActive}
+                                                toggleContextMenuFromActiveReportAction={toggleContextMenuFromActiveReportAction}
+                                                handleShowContextMenu={handleShowContextMenu}
+                                                showPopover={showPopover}
+                                                resolveActionableMentionWhisper={resolveActionableMentionWhisper}
+                                                resolveActionableReportMentionWhisper={resolveActionableReportMentionWhisper}
+                                                currentSearchHash={currentSearchHash}
+                                            />
+                                        </ReportActionItemFrame>
                                     </SearchActionHeader>
                                 </OfflineWithFeedback>
                             </View>
